+++ packages/tailwind-preset/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1-next.14](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@1.1.1-next.13...@esri/calcite-tailwind-preset@1.1.1-next.14) (2026-05-22)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [1.1.1-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@1.1.1-next.12...@esri/calcite-tailwind-preset@1.1.1-next.13) (2026-05-22)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
 ## [1.1.1-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@1.1.1-next.11...@esri/calcite-tailwind-preset@1.1.1-next.12) (2026-04-28)
 
 **Note:** Version bump only for package @esri/calcite-tailwind-preset